+++ src/platform_impl/windows/window.rs
@@ -98,6 +98,11 @@ impl Window {
     }
 
     #[inline]
+    pub fn is_visible(&self) -> Option<bool> {
+        Some(unsafe { winuser::IsWindowVisible(self.window.0) == 1 })
+    }
+
+    #[inline]
     pub fn request_redraw(&self) {
         unsafe {
             winuser::RedrawWindow(